# packages/server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger",
    "description": "KVAS Forum API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1"
    }
  ],
...
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get current user",
        "description": "Get user",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ID": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 100
          }
        }
      },
...
      "Avatar": {
        "required": [
          "avatar"
        ],
        "type": "object",
        "properties": {
          "avatar": {
            "type": "string",
            "example": "/path/to/avatar.jpg"
          }
        }
      },
...
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ID"
          },
          {
            "$ref": "#/components/schemas/Avatar"
          },
          {
            "required": [
              "user"
            ],
            "type": "object",
            "properties": {
              "first_name": {
                "type": "string",
                "example": "Petya"
              },
              "second_name": {
                "type": "string",
                "example": "Pupkin"
              },
              "display_name": {
                "type": "string",
                "example": "Petya Pupkin"
              },
              "login": {
                "type": "string",
                "example": "userLogin"
              },
              "email": {
                "type": "string",
                "example": "my@email.com"
              },
              "phone": {
                "type": "string",
                "example": 89223332211
              },
              "role": {
                "type": "string",
                "example": "regular"
              }
            }
          }
        ]
      },
...
      "Unauthorized": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
...
      "ServerError": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Internal Server Error"
          }
        }
      }
    }
  }
}